% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreMRDQ.R
\name{scoreMRDQ}
\alias{scoreMRDQ}
\title{Score the Michigan Retinal Degeneration Questionnaire}
\usage{
scoreMRDQ(dat, verbose = 0)
}
\arguments{
\item{dat}{data.frame with N rows and 60 columns.  First column is patient ID and the remaining columns are responses to the 59 items.  Responses should be integers from 0 to 5.}

\item{verbose}{A number indicating the amount of printing during function execution. 0 (default) is none. Higher numbers may result in more printing.}
}
\value{
A list with two components, each an N by 7 matrix. The first, named thetas, contains the 7 domain scores for each patient.  The second, named ses, contains the standard errors of the thetas.
}
\description{
Score the Michigan Retinal Degeneration Questionnaire
}
\examples{
justme <- read.table(
  text = "Me,0,0,0,0,0,0,0,0,1,1,1,0,0,0,1,1,1,2,0,0,0,0,1,1,1,0,0,0,0,1,0,0,1,0,0,1,0,0,1,0,0,1,0,0,1,0,0,2,1,1,0,0,0,0,0,2,0,0,0",
  sep = ",")
scoreMRDQ(justme)
}
